102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,16 +6,16 @@
  */
 int main(void)
 {
-long long int n1 = 1, n2 = 2, n3;
+long int n1 = 1, n2 = 2, n3;
 
 int i, n = 50;
-printf("%lld, %lld, ", n1, n2);
+printf("%ld, %ld, ", n1, n2);
 for (i = 3; i <= n; ++i)
 {
 
 n3 = n1 + n2;
 
-printf("%lld, ", n3);
+printf("%ld, ", n3);
 n1 = n2;
 n2 = n3;
